Convert Textures.rsrc to TGA
The resources were PICTs, but it's easier to work with TGA.

The raw RGB data in the converted TGAs is a straight copy from the
PICTs' uncompressed RGB data, so this conversion is lossless.

As a result, the original 16-bit 5-5-5 color format of most of the files
was preserved. Only 202.tga, 1001.tga and 1002.tga are 24-bit RGB like
their PICT counterparts.