Revert pixel switching and fix texture bugs
Before, I had swapped the pixel order in Bugdom and Pomme to work around
the VitaGL implementation. Now, I have instead implemented the color
order in VitaGL and have reverted it in Bugdom and Pomme.
This should siimplify maintenance in the future.

While implementing the pixel conversion in VitaGL, I fixed the 1-byte
offset texture bug that was making everything look grainy.