Framerate-independent damage logic in lv8 boulders (#38)
Whether the boulders would yield damage or not used to depend on the framerate. The OS 9 original had an uncapped framerate, so, there's no "blessed" reference for this behavior.

But, it seems reasonable to assume that the original speed floor of 10 world units per delta was calibrated for a framerate in the ballpark of 30 fps.

So, the chosen damage floor for this fix (0.1) works out to a delta of 10 units at 27 fps.